requirements.txt
numpy
Pillow
matplotlib
torch
torchvision
